Cast returned Integers from DB as Longs...because of Anko...

+++ app/src/main/kotlin/com/om/snipit/models/Book.kt
@@ -1,5 +1,5 @@
 package com.om.snipit.models
 
-data class Book(val id: Int = -1, val title: String, val author: String, val image_path: String,
+data class Book(val id: Long = -1, val title: String, val author: String, val image_path: String,
     val date_added: String,
-    val color_code: Int, val list_order: Int)
+    val color_code: Long, val list_order: Long)
